Split the largest form components and unify the remount keys
