+++ src/components/Routines/widgets/forms/ProgressionForm.tsx
@@ -5,7 +5,7 @@ import { Box, Button, Divider, IconButton, MenuItem, Stack, Switch, TextField, T
 import Grid from "@mui/material/Grid";
 import Tooltip from "@mui/material/Tooltip";
 import { useAppForm } from "@/core/forms/appForm";
-import { yupSchema, submitHandler } from "@/core/forms/formUtils";
+import { defaultsKey, submitHandler, yupSchema } from "@/core/forms/formUtils";
 import { FormQueryErrors } from "@/core/ui/Widgets/FormError";
 import {
     BaseConfig,
@@ -13,17 +13,21 @@ import {
     OPERATION_REPLACE,
     OPERATION_VALUES_SELECT,
     OperationType,
-    REQUIREMENTS_VALUES,
     STEP_VALUES_SELECT,
     StepType
 } from "@/components/Routines/models/BaseConfig";
 import { useProcessConfigsQuery } from "@/components/Routines/queries/configs";
 import { ConfigDetailsRequirementsField, ConfigType } from "@/components/Routines/widgets/forms/BaseConfigForm";
+import {
+    emptyEntry,
+    progressionEntries,
+    progressionPayload,
+    progressionSchema,
+    ProgressionFormValues
+} from "@/components/Routines/widgets/forms/progressionFormData";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { AddBaseConfigParams, EditBaseConfigParams } from "@/components/Routines/api/baseConfig";
 import { ApiPath } from "@/core/lib/consts";
-import * as yup from "yup";
 
 interface ProgressionFormProps {
     configs: BaseConfig[],
@@ -36,66 +40,14 @@ interface ProgressionFormProps {
     isWeeklyCycle: boolean;
 }
 
-interface ProgressionFormValues {
-    entries: BaseConfigEntryForm[],
-}
-
-const getEmptyConfig = (iter: number, edited: boolean, forceInteger: boolean): BaseConfigEntryForm => ({
-    forceInteger: forceInteger,
-
-    edited: edited,
-    iteration: iter,
-
-    id: null,
-    idMax: null,
-    value: '',
-    valueMax: '',
-    operation: OPERATION_REPLACE,
-    operationMax: OPERATION_REPLACE,
-    step: "abs",
-    stepMax: "abs",
-    requirements: [],
-    requirementsMax: [],
-    repeat: false,
-    repeatMax: false,
-});
-
 export const ProgressionForm = (props: ProgressionFormProps) => {
     const forceInteger = props.forceInteger ?? false;
-
-    const defaultEntries: BaseConfigEntryForm[] = [];
-    for (const iteration of props.iterations) {
-        const config: BaseConfig | undefined = props.configs.find((c) => c.iteration === iteration);
-        const configMax: BaseConfig | undefined = props.configsMax.find((c) => c.iteration === iteration);
-
-        if (config === undefined) {
-            defaultEntries.push(getEmptyConfig(iteration, false, forceInteger));
-        } else {
-            defaultEntries.push({
-                forceInteger: forceInteger,
-
-                edited: true,
-                id: config.id,
-                idMax: configMax === undefined ? null : configMax.id,
-                iteration: iteration,
-                value: String(config.value),
-                valueMax: configMax === undefined ? '' : String(configMax.value),
-                operation: config.operation,
-                operationMax: configMax === undefined ? OPERATION_REPLACE : config.operation,
-                step: config.step,
-                stepMax: configMax === undefined ? "abs" : configMax.step,
-                requirements: config.requirements?.rules ?? [],
-                requirementsMax: configMax === undefined ? [] : configMax.requirements?.rules ?? [],
-                repeat: config.repeat,
-                repeatMax: configMax === undefined ? false : config.repeat,
-            });
-        }
-    }
+    const defaultEntries = progressionEntries(props.iterations, props.configs, props.configsMax, forceInteger);
 
     // The form freezes its default values, so saved or reloaded configs get a
     // fresh form via the key
     return <ProgressionFields
-        key={JSON.stringify(defaultEntries)}
+        key={defaultsKey(defaultEntries)}
         {...props}
         forceInteger={forceInteger}
         defaultEntries={defaultEntries}
@@ -142,182 +94,17 @@ const ProgressionFields = (props: ProgressionFormProps & {
             break;
     }
 
-    const validationSchema = yup.object({
-        entries: yup.array().of(
-            yup.object().shape({
-                edited: yup.boolean(),
-                iteration: yup.number().required(),
-
-                // Conditionally apply integer validation e.g. for sets
-                value: yup.number()
-                    .when('forceInteger', {
-                        is: true,
-                        then: schema => schema.integer(t('forms.enterInteger')).typeError(t('forms.enterNumber')),
-                        otherwise: schema => schema.typeError(t('forms.enterNumber')).nullable().notRequired(),
-                    }),
-
-                // only check that the max number is higher when replacing. In other cases allow the max
-                // weight to e.g. increase less than the min weight
-                valueMax: yup.number().typeError(t('forms.enterNumber')).nullable()
-                    .when('forceInteger', {
-                        is: true,
-                        then: schema => schema.integer(t('forms.enterInteger')).typeError(t('forms.enterNumber')),
-                        otherwise: schema => schema.typeError(t('forms.enterNumber')).nullable().notRequired(),
-                    })
-                    // Conditionally apply integer validation e.g. for sets
-                    .when('operation', {
-                        is: OPERATION_REPLACE,
-                        then: schema => schema.min(yup.ref('value'), t('forms.maxLessThanMin')),
-                        otherwise: schema => schema,
-                    }),
-                operation: yup.string().required(),
-                operationMax: yup.string().required(),
-                requirements: yup.array().of(yup.string().oneOf(REQUIREMENTS_VALUES)),
-                requirementsMax: yup.array().of(yup.string().oneOf(REQUIREMENTS_VALUES)),
-                repeat: yup.boolean(),
-                repeatMax: yup.boolean()
-            })
-        )
-            .test(
-                'inter-entry-validation',
-                'Your error message here',
-                function (entries) { // Use 'function' to access 'this'
-                    const { createError } = this;
-
-                    const data = entries as unknown as BaseConfigEntryForm[];
-
-                    for (let i = 0; i < data.length; i++) {
-                        const entry = data[i];
-
-                        // If there is an entry down the line
-                        // if (entry.iteration === 1 && data.length > 1 && !entry.value && entry.edited) {
-                        //     return createError({
-                        //         path: `entries[${i}].value`,
-                        //         message: 'Value is required at workout nr 1 when other entries exist'
-                        //     });
-                        // }
-
-                        if (entry.iteration > 1 && entry.operation !== OPERATION_REPLACE) {
-                            let hasValuePreviousReplace = false;
-                            let hasMaxValuePreviousReplace = false;
-
-                            for (let j = 0; j < i; j++) {
-                                if (data[j].operation === OPERATION_REPLACE && data[j].value !== '' && data[j].edited) {
-                                    hasValuePreviousReplace = true;
-                                }
-
-                                if (data[j].operation === OPERATION_REPLACE && data[j].valueMax !== '' && data[j].edited) {
-                                    hasMaxValuePreviousReplace = true;
-                                }
-                            }
-
-                            if (!hasValuePreviousReplace) {
-                                return createError({
-                                    path: `entries[${i}].value`,
-                                    message: t('routines.progressionNeedsReplace')
-                                });
-                            }
-                            if (!hasMaxValuePreviousReplace) {
-                                return createError({
-                                    path: `entries[${i}].valueMax`,
-                                    message: t('routines.progressionNeedsReplace')
-                                });
-                            }
-                        }
-                    }
-
-                    // All entries valid
-                    return true;
-                }
-            )
-        ,
-    });
-
-    const handleSubmit = (values: ProgressionFormValues) => {
-        // Remove empty entries
-        const data = values.entries.filter(e => e.edited);
-
-        // Split between min and max values
-        const editList: EditBaseConfigParams[] = data.filter(data => data.id !== null).map(data => ({
-            id: data.id!,
-            // eslint-disable-next-line camelcase
-            slot_entry: props.slotEntryId,
-            value: data.value as number,
-            iteration: data.iteration,
-            operation: data.operation,
-            step: data.step,
-            repeat: data.repeat,
-            requirements: { rules: data.requirements ?? [] }
-        }));
-        const addList: AddBaseConfigParams[] = data.filter(data => data.id === null && data.value !== '').map(data => ({
-            // eslint-disable-next-line camelcase
-            slot_entry: props.slotEntryId,
-            value: data.value as number,
-            iteration: data.iteration,
-            operation: data.operation,
-            step: data.step,
-            repeat: data.repeat,
-            requirements: { rules: data.requirements ?? [] }
-        }));
-        // Items to delete, also includes all where the value is empty
-        const deleteList = props.configs.filter(c => iterationsToDelete.includes(c.iteration)).map(c => c.id);
-        data.forEach(entry => {
-            if (entry.value === "" && entry.id !== null && !deleteList.includes(entry.id)) {
-                deleteList.push(entry.id);
-            }
-        });
-
-        // Max values
-        const editListMax: EditBaseConfigParams[] = data.filter(data => data.idMax !== null && data.valueMax !== '').map(data => ({
-            id: data.idMax!,
-            // eslint-disable-next-line camelcase
-            slot_entry: props.slotEntryId,
-            value: data.valueMax as number,
-            iteration: data.iteration,
-            operation: data.operation,
-            step: data.step,
-            repeat: data.repeat,
-            requirements: { rules: data.requirements ?? [] }
-        }));
-        const addListMax: AddBaseConfigParams[] = data.filter(data => data.idMax === null && data.valueMax !== '').map(data => ({
-            iteration: data.iteration,
-            // eslint-disable-next-line camelcase
-            slot_entry: props.slotEntryId,
-            value: data.valueMax as number,
-            operation: data.operation,
-            step: data.stepMax,
-            repeat: data.repeat,
-            requirements: { rules: data.requirements ?? [] }
-        }));
-        // Items to delete, also includes all where the value is empty
-        const deleteListMax = props.configsMax.filter(c => iterationsToDelete.includes(c.iteration)).map(c => c.id);
-        data.forEach(entry => {
-            if (entry.valueMax === "" && entry.idMax !== null && !deleteList.includes(entry.idMax)) {
-                deleteListMax.push(entry.idMax);
-            }
-        });
-
-        // Save to server
-        processEntriesQuery.mutate({
-            values: {
-                toAdd: addList,
-                toDelete: deleteList,
-                toEdit: editList,
-                apiPath: apiPath
-            },
-            maxValues: {
-                toAdd: addListMax,
-                toDelete: deleteListMax,
-                toEdit: editListMax,
-                apiPath: apiPathMax
-            }
-        });
-    };
-
     const form = useAppForm({
         defaultValues: { entries: props.defaultEntries } as ProgressionFormValues,
-        validators: { onChange: yupSchema<ProgressionFormValues>(validationSchema) },
-        onSubmit: async ({ value }) => handleSubmit(value),
+        validators: { onChange: yupSchema<ProgressionFormValues>(progressionSchema(t)) },
+        onSubmit: async ({ value }) => processEntriesQuery.mutate(progressionPayload(value.entries, {
+            slotEntryId: props.slotEntryId,
+            configs: props.configs,
+            configsMax: props.configsMax,
+            iterationsToDelete,
+            apiPath,
+            apiPathMax,
+        })),
     });
 
     return <>
@@ -383,12 +170,12 @@ const ProgressionFields = (props: ProgressionFormProps & {
                                                 if (log.id !== null) {
                                                     setIterationsToDelete([...iterationsToDelete, log.iteration]);
                                                 }
-                                                form.replaceFieldValue('entries', index, getEmptyConfig(log.iteration, false, forceInteger));
+                                                form.replaceFieldValue('entries', index, emptyEntry(log.iteration, false, forceInteger));
                                             }}>
                                             <DeleteIcon />
                                         </IconButton>
                                         : <IconButton size="small" onClick={() => {
-                                            form.replaceFieldValue('entries', index, getEmptyConfig(log.iteration, true, forceInteger));
+                                            form.replaceFieldValue('entries', index, emptyEntry(log.iteration, true, forceInteger));
                                         }}>
                                             <AddIcon />
                                         </IconButton>
@@ -399,7 +186,8 @@ const ProgressionFields = (props: ProgressionFormProps & {
                                 <Grid size={2}>
                                     {log.edited &&
                                         <form.AppField name={`entries[${index}].value`}>
-                                            {field => <field.WgerTextField variant="standard"
+                                            {field => <field.WgerTextField
+                                                variant="standard"
                                                 title={t('min')}
                                                 fullwidth={true}
                                                 fieldProps={{ slotProps: { htmlInput: { inputMode: 'decimal' } } }}
@@ -409,7 +197,8 @@ const ProgressionFields = (props: ProgressionFormProps & {
                                 <Grid size={2}>
                                     {log.edited &&
                                         <form.AppField name={`entries[${index}].valueMax`}>
-                                            {field => <field.WgerTextField variant="standard"
+                                            {field => <field.WgerTextField
+                                                variant="standard"
                                                 title={t('max')}
                                                 fullwidth={true}
                                                 fieldProps={{ slotProps: { htmlInput: { inputMode: 'decimal' } } }}
