+++ src/core/forms/formContexts.ts
@@ -2,4 +2,4 @@ import { createFormHookContexts } from "@tanstack/react-form";
 
 // Separate from appForm.ts so the bound field components can import the
 // contexts without importing the hook that registers them
-export const { fieldContext, formContext, useFieldContext, useFormContext } = createFormHookContexts();
+export const { fieldContext, formContext, useFieldContext } = createFormHookContexts();
